Refactored Factory
I refactored the factory to grab the parameter information when we are getting which valitators to apply. I also made sure that we would be able to deserialize directly to a JObject when deserializing our options object. I will port this code over to the real project on Monday. Sorry about not getting it done this week :(.